editor: bring back "Reveal in Project Tree"
Disabled this during the plugins extraction work, can bring it back now
and move the helper to GbWorkbench.

Long term, I'd like to figure out a clean way to move the project tree to
plugins, which will complicate this a bit. However, once we have a
GbEditorViewAddin (or whatever better naming we come up with), the
project-tree could provide this feature itself.