search: refactor search into PeasExtension
This moves the devhelp search provider to be implemented as a peas
extension. We had to change the interfaces a bit to be able to do this,
but should be much easier to bridge the libide<->ui stuff going forward.

We still need to update the git index as well, but we don't want that to
be git based going forward (it needs to crawl the filesystem) so that it
works with regular directory trees too.