tree: make GbTree build nodes lazily
This changes GbTree to only build tree nodes lazily. This means a couple
things for consumers. For example, performing a foreach on the tree
will no longer yield results unless the item has already been added.

Such features such as "Reveal in Tree" will need to be updated to deal
with this.

However, what this lets us do is change things like the Project Tree to
no longer use IdeProjectFiles, but instead lookup files from the file-
system directly, and lazily. Doing so should improve support for larger
project trees greatly.

Still need to deal with the fallout from this, but this covers the
plumbing required.