workbench: make gb_workbench_reveal_file() work with lazy tree building
Now that not all the nodes have been built, we need to be more precise
with our lookups. This uses the new gb_tree_find_child_node() helper.

As a side-effect, we don't have to scan the entire tree anymore either.

The gb_tree_find_child_node() function will ensure the node is built
before scaning the children. So tree paths following the lookup request
will be opportunistically built.