clang: add minimal implementation of symbol tree for Clang
Lots still to be done here. For example, we need to support more clang
cursor types. We also need to create source locations for the nodes so
that we can jump to the location when activated in the tree.

We should make sure we do source location lookup via a vfunc/signal so
that it can be done lazily as well. No sense in creating all those
file path strings until they are needed.