plugins: switch to using dynamic shared libraries for plugins
We might change back to embedded if/when it lands in libpeas, but this
lets us move forward now and start merging back into master.

We should also start cleaning up the plugin code, which can probably be
simplified now that I have a hang of how this is getting put together.