tree: add support for GbTreeNode:use-dim-label
This gproperty will result in the text of the node being rendered with
a dim label. We need to manually set the style for various themes since
the alpha value is not extracted from the style context (always set to
1.0). Possibly a better way to deal with that.

I'm unsure if we should also apply this to the pixbuf renderer.