project-tree: use strcmp instead of g_utf8_collate to compare collation keys
It seems it is possible for g_utf8_collate_key_for_filename() to return an
invalid UTF-8 string, which causes crashes in g_utf8_collate() because the
string cannot be normalized. According to the documentation of
g_utf8_collate_key_for_filename(), we should just use strcmp() to compare
its return value.

https://bugzilla.gnome.org/show_bug.cgi?id=752265