python: add very basic auto-completion for python using jedi
Long term, this really should be put in gnome-code-assistance, but we
still need to figure out the right abstraction there. Creating
abstractions before we understand what is needed is generally a bad idea,
so lets prototype this in process.

Igor Gnatenko mentioned that Jedi is a pretty good completion library for
Python. Thankfully, it seems to be packaged on a few systems including
Fedora.

If jedi is not found, this plugin will simply do nothing. On Fedora
systems, "dnf install python3-jedi" should get the job done.

Currently, this performs the lookup using Python's big Thread. I don't
like this because it means we get all the python memory fragmentation
with it. This is another great reason for this to belong in
gnome-code-assistance. Multiprocessing is another short term stop gap
that could be used. The threading/GIL shouldn't affect us much here,
but with CPython, always famous last words.

If you know python, I'm sure you can improve this module.

Patches wanted, accepted.