ctags: add symbol resolver using ctags
This adds support for jumping to symbols by extracting the location using
the ctags index. The ctags index may contain a Regex, so we might have to
apply the regex against the mapped file (or loaded buffer) from a thread.

After resolving that, we can notify back to the UI thread where the
discovered location was.

Alternatively, ctags might contain a line number we can use and avoid the
whole process. I've only seen that with C files so far.

Tested with python and ruby, and both seem to work, since ctags heavily
uses regexes to define the location (such as /^   def foo(self):$/).

We do escape () to be '.', but I don't think that will cause too much
problems. Alternatively we should fix the regex to be more literal.

Line number support coming forthwith.