snippets: store snippet text with snippet
This can be handy to have so that we can visualize the actual snippet text
to the user in various places (snippet editor for example).

This is definitely complicated by the single snippet reuse in multiple
languages. But I think if we continue down this path and track where,
in which file, the snippet originated, we can point users to the correct
location to edit them.