build: make libgnome-builder.la install to private directory
In preparation for allowing plugins to get access to introspection data
for the Builder process, we want to make libgnome-builder.la an installable
(but private) shared library. We'll place it in the same place as our
other private libraries ($libdir/gnome-builder).

This also makes libegg and librg available as a shared library so we can
more easily link against it from various libraries.