jedi: use IdeSourceSnippet for jedi completions
This allows us to complete functions with the function names as tab stops.
We still need to improve the GObject Introspection so that function calls
actually appear as "function".

Functions, (at least those defined in python for now), will be expanded
like:

  self.foo(a, b, c)

and you can tab through a, b, and c. A final tab will land you after the
trailing ).