jedi: improve interactivity by tracking cancellation
We still have more to do here in terms of making this interactive.
Python really blocks us at some inopportune moments. Getting this out
of process will also likely help. To do that, we need to finish the
autocompletion API for G-C-A and do the same wrapping trick we
do for Clang completion (but with GVariant).