view-stack: use GtkPopover and GMenuModel for view stack menu
This is not an ideal thing, as I think our custom design before looked
better than what gets generated by gtk_popover_bind_model(). However, we
really need to move in this direction so that plugins can merge menu items
with the system.

I think this design is going to change in the near future, so probably an
okay change to get where we are going.

C'est la vie