hack: delay GtkTextView size-allocate during spurious animation
This hack is to delay size-allocate in GtkTextView until after the
animation sequence has completed. Currently, this hack only works for
shrinking the animation. Such occurrence happens when the panels
animate in.

We don't really need to re-expose the content until after the sequence
completes. This is potentially invalid for RTL or right-aligned text
during the sequence, but it is not totally unexpected for that content
to "snap" into place after a live resize. (See Safari on OS X for an
example of how they do this with 'inLiveResize').

It would be much better if we could set if we are in a live resize on
a GdkWindow so that animation sequences and Drag'n'Resize could be
propagated to widgets to allow them to "do less work" during the draw
cycle. GtkTextView might choose to work extra hard to not drop it's
pixel cache in this case.

The result here, is that you can animate all three panels in smoothly
now. (Animating out is still a big jagedy since we don't mitigate that
yet). Some hardware, of course, may vary.