theme: tidy up notebook/header styling on Adwaita
You'll notice the editor tweak widget is not matching here. This is because
it is getting clipped so the box shadow won't show up. What I'd like to
do to fix this is to pull the editor tweak out of the GbEditorView and into
a plugin that attaches itself to the view stack. It will enable/disable
itself based on view type. This will allow us to get rid of the
"controls" stack too.