egg-counter: add vsdo fallback when on linux and no rdtscp instruction
This is basically as fast as rdtsp but more portable since it will work
with Linux even on machines without rdtscp. If for some reason we don't
have any of that (ARM Linux) we fallback to sched_getcpu().

BSDs should continue to use atomics fallback. I would love to see that
change though.