editor: add simple goto line popover
I don't really like that we use a popover per view for this, but at least
we can say we have this now. We can optimize the widget overhead later.

We should also go the extra mile to check that the line exists while
typing them.

We also need keyboard shortcuts.