credits: update application credits
Those in the funders list donated for a perk via indiegogo at the $25 tier
or higher. If you are not in this list, and expect to be, it is probably
because your donation was marked as anonymous. Please send a patch or ping
me at christian@hergert.me to have this corrected.

Thanks everyone!

Signed-off-by: Christian Hergert <christian@hergert.me>