clang: make clang autocompletion fast
Okay, so there is some dark wizardry going on here.

There are two major shitstorms to optimize around here. The first, is that
clang will basically give us everything in the world back that could be
completed, regardless of if it is actually useful. For example, while
writing this module, I was seeing an average of 30,000 entries. Of course,
the are sorted in no particular order.

The second shitstorm, is that GtkSourceCompletionContext takes a GList
of items. I'm not particularly keen on allocating lots of GList items
especially if it means we need to sort them.

So instead, we have a double structure where we keep an array of
completion items, of which embed a GList node in them. This way, we
don't need to allocate any GList items at all, and have a single large
contiguous array for each of our completion items.

When drilling down in search, it makes sense to not requery any of the
previous items. This is where that embedded GList really saves the
day again. Instead of traversing the GPtrArray for all the matching items,
we simply research the GList starting from our synthesized head. Therefore,
each key press continually reduces N, the number of items to refilter.

Upon each request to populate the completion set, we check to see if
we can reuse the existing data set. This is done by looking backwards at
the line and seeing if any character that could invalidate the set has
been typed. If so, we go through the refilter step above.

If we backspace, we won't match our linked list optimization, so we have
to reset the linked list and refilter. This could be optimized a bit more,
but we'll see if it is really an issue first. (To optimize it, avoid the
rebuild of the linked list be fore rewalking the list to filter).

This still doesn't sort items by their priority. We need to come up
with a decent algorithm for determining that in a quick fashion, as
we probably want to sort the whole list by that hueristic up front.

Enjoy!