libide: add IdeCompletionItem and IdeCompletionResults helpers
These abstractions make it easier to write fast autocompletion providers
for GtkSourceView. They abstract the GList embedded link node hack as well
as dealing with checking if a query can be replayed (and performing that
as well).

Consumers should override IdeCompletionItem::match() and
IdeCompletionResults::compare() (or set the items priority appropriately).