workbench: don't use property for workbench in GbWorkbenchAddin
Using properties is much more pain than it is worth. Instead, just pass
the workbench during the load and unload vfuncs like other addins do.

Also, we need to change plugins that are adding "themselves" to the UI,
as using libpeas+GInitiallyUnowned seems to be rather bad from an object
lifecycle standpoint. It also makes GtkWidget::destroy() a pain to
work with.