libide: only replay results if we have symbol characters
We only use this for symbols right now, so safe to put this check here.

The gist is this, if we have a character like ( or [ or {, our result
set would not take that into account. So callers need to generate a new
list. We simply limit this to "symbol chars" which are alpha numeric
today. Theoretically, we might need to support ? and ! at some point,
but those are always suffixes and require us to change the result set
anyway.