python: start on a new python-based indenter for python
This is the beginning of something significantly more advanced than our
current indenter. However, it still has a lot of work to be done.

The gist is this, we first do a pass of ranking the various scope affecting
syntax items in our buffer. Then we sort those rankings, and possibly
prune the tree a bit.

After that, we can do some fairly interesting lookups and apply rules
for them. In particular, it should become possible to format function
parameters and function calls and tuples all differently.

Discovering a function declaration would look similar to the following:

    match = discoveries.select(Rank.FUNCTION, Rank.TUPLE)
    if match and match[0].line == match[1].line:
        ...