libide: add IdeTagsBuilder
We need a basic tags building interface that various subsystems can
implement to delegate an implementation. For example, the ctags service
can build ctags, but so can autotools with `make ctags`. And we'd like
to prefer `make ctags` since it is a lot faster by avoiding building
when dependencies haven't changed.