ide-source-view: add save-command css binding
We can now use "save-command" in css binding to remember the current
command so we can get it later in priv->command.
It's required by Vim mode when some commands rely on contextual
informations from a previous command.

It works by recording the current GdkEvent pressed key so you need
to put it at the start of your binding to work.

The command is also passed to _ide_source_view_apply_movement ()
movements dispatcher.