egg: add EggSettingsFlagAction
This class provides a GAction that can toggle a single flag in a GSettings
key of the "flags" type. "Flags" gsettings are really just an array of
string key nicknames. This will toggle on/off those strings by first
doing a readback, adding the string, and the writing the new value.
Obviously that is racey, but who does settings on threads...

The action name will be the schema-key and flag nickname separated by a
minus (-). An example for "draw-spaces" key and "space" nickname would
be "draw-spaces-space".

This can be connected to a GtkCheckButton relatively easily by setting
the action name to the example name above.