prefs: add options to toggle drawing spaces within the editor
This gives some, long missing, checkboxes to toggle flags within GSettings.
I mostly delayed on this before because I didn't have a clean way to do
this. Not that this is super clean, but it is something!

Translators, I'm sorry for putting this in so late in the 3.18.1 cycle,
but I'd rather have it than not when 3.18.1 will most likely be our last
3.18 release.