tree: implement recursive search for tree nodes
This will only filter items that have been loaded. If you have a tree that
needs to dynamically load data, you are expected to have done that, or
be doing that, in the background iteratively.

This will match the filters as items are added or already existing in
the underlying TreeStore.