pygobject: add Builder.py overrides
This allows us to create a GDBus-based service without all of the pain
of multiple dbus implementations in process (python-dbus and GDBus). It
only supports methods currently, and is based on a proposal from Martin
Pitt which can be found on

  https://bugzilla.gnome.org/show_bug.cgi?id=656330

This can be used by worker processes to communicate with the parent
Builder process over a soon to be added private bus.