libide: add IdeWorker interface and plumbing
IdeWorker is designed to allow for a plugin to create a worker process
where it will perform it's heavy work. Plugins that could potentially
consume a lot of memory or resources should be pushed into a worker
process so that we can recycle them based on future policies.

The worker manager creates a GDBusServer that can be connected to by the
worker process. It should export it's service via that bus, and connect
to it from the UI process (using API to be added soon).

The UI should handle for the potential that the connection and proxy are
lost in the case that the worker process crashes, or is torn down due to
overactive memory/cpu use.