jedi: perform autocompletion lookup in a subprocess
This uses the new IdeWorker API to create a worker and perform the
autocompletion in that worker process. The results are then passed back
to the UI process via GDBus, and rendered in the completion window.

There is of course a performance penalty for this, but it helps reduce
fragmentation in the parent process and we can incrementally improve
on this. I'd like to see the children processes recycled occasionally.

Also, the worker process is spawning too many threads today, but that
can be improved upon in followup commits.