prefs: start on preferences perspective
The goal of this is to allow for editing preferences from the workbench
window without having to open a transient dialog. While we have an
IdePreferences interface that can be implemented for a dialog as well,
I doubt we will do that as this seems much more convenient.

We might consider embedding the persepctive in a dialog for a case where
the user just wants to edit preferences and not run Builder, but not too
worried about that case.