jedi: show documentation next to completion suggestions
This is a prototype to get documentation from .gir files and show it
next to our completion suggestions. It uses a simple sqlite db for
storing the documentation, and adds a dependency on lxml for parsing
the documentation.

For the long term, it would be a good idea to port this to C so other
plugins (eg. the clang one) could use it to show documentation too.

I hope we can use this work to later add a context menu item
"show documention" to get documentation for the selected symbol.