prefs: push schema-id and path into container for expansion
We intend to have sub-pages which can be pushed onto the preferences stack.
This requires that we be able to expand some simple key/value pairs so that
the transient page can be based on user input (such as a source language
id). Pushing the expansion code into container seems like the natural
choice.

We still need to add transient pages and API for setting that key/value
map.