source-view: Use bottom margin to scroll beyond the last line
The overlay scrollbars can cover the last line which makes it
difficult to view and select it properly. These changes add a
bottom margin to allow scrolling beyond the last line.

This bottom margin can be set by using the property "overscroll"
which receives the number of lines to scroll beyond the end of the
buffer. If a negative number is given, it counts the lines from the
top of the viewport (instead of the bottom).