libide: start on very basic GMenu merging for plugins
The goal here is to allow plugins to simply provide a gtk/menus.ui like
an application would. However, the menus will get merged into the real
menus within Builder.

For example, we could extend the context menu for the source view with
something like:

<?xml version="1.0"?>
<interface>
  <menu id="ide-source-view-popup-menu">
    <section id="ide-source-view-popup-menu-reveal-file-section">
      <attribute name="after">ide-source-view-popup-menu-selection-section</attribute>
      <item>
        <attribute name="label" translatable="yes">Re_veal in Project Tree</attribute>
        <attribute name="action">project-tree.reveal-file</attribute>
      </item>
    </section>
  </menu>
</interface>

The interesting part here is to implement "after" and "before" properly
as well as support more than just sections (as we do now).