build-result: add IdeBuildResult::log signal
This adds a signal that is emitted when a new log line is received from
stderr or stdout. I'm not hugely fond of this design (since signals are
pretty lock-heavy), but I rather like the API. It makes it very simple
for a build result observer to track things. We'll always be in the main
loop anyway, so contention should be low.

We should be able to use this signal to move errorformat extraction out
of IdeBuildResult too.