gcc: add gcc plugin to perform error-format extraction during builds
This moves the error extraction for diagnostics to a plugin when using GCC
in automake. It should work for other build systems that log, given that
they spit gcc diagnostics to stdout/stderr.

Ideally, in the future, this plugin would include things that would make
us not dependent on clang.