contributing: start on contributing bootstrap
This is just the sketch of an idea for a contributing plugin. The idea is
a few fold. First, we want to simplify the process of going from nothing
to something as quick as possible. While we are focused on bootstraping
GNOME projects, I don't see why this can't be used in a broader sense.

This is just a minimal amount of code to get started. I'd love for someone
to work on making this look more like the description in the IDEA file.

We also need to figure out how we are going to hook into jhbuild/xdg-app
to take care of ensuring we have necessary project dependencies.

We also need to either extend DOAP or extract PKG_CHECK_MODULES() from
configure.ac. We could also rely on modulesets to some degree.