plugins: Make the symbol-tree a GObject instead of a GtkWidget
The ref counting of this extension being a GtkWidget does not work, as
expected, causing crashes whenever it gets disabled since it expects a
GObject and not a floating object like the GtkBox it inherits from.

To fix this issue, this patch onverts the extension into a GObject and
extract the UI code to a new class SymbolTreePanel (subclass of GtkBox).

https://bugzilla.gnome.org/show_bug.cgi?id=761106