library-template: Port from glib-gettext/intltool to upstream gettext
This requires version 0.19.6 of upstream gettext for AppData support:

https://lists.gnu.org/archive/html/info-gnu/2015-09/msg00003.html

It is an error to use glib-gettext and intltool together (as the
template was previously doing) — either use one or the other. However,
both are deprecated in favour of upstream gettext, which is alive again:

https://wiki.gnome.org/Projects/GnomeCommon/Migration#line-204

This has not been extensively tested, but should be the right general
shape.

https://bugzilla.gnome.org/show_bug.cgi?id=761356