egg: add EggColumnLayout
EggColumnLayout is a widget that will try to layout children widgets in
a column, but overflow to horizontal columns as space is available.

This has the property that it will try to flow horizontally before
vertically, but if vertical flowing is necessary, then the widget will
still preserve the ordering of the children.