gnome-builder: Set tooltips on header bar buttons
It is an enhancement of the gnome-builder app which intends to inform the users by adding tooltips to buttons.

Now the users will know exactly what each button does before clicking it.

To fix this i edited some files and added the tooltip-text with translatable property on : The search bar, the menu button, the toggle button for left, bottom, right panels and also the Search bar on the Preferences perspective.

https://bugzilla.gnome.org/show_bug.cgi?id=755185