libide: use setsid() and kill(-pid, SIGKILL) to kill process group
We had an issue where "gmake", despite getting SIGKILL, would not cause
the build process to exit. This seems to be caused by the descendent
processes not dying with the parent.

To allow us to kill all of the processes, we first create a new session
group with setsid() in our child setup function. If the process should
be cancelled (via g_cancellable_cancel()), then we send SIGKILL to all
members of the session group with kill(-pid, SIGKILL). Specifying a
negated(pid) indicates delivery to all members of the group.

We probably need to verify that this works on FreeBSD, but it seems
plausible that will.

https://bugzilla.gnome.org/show_bug.cgi?id=760874