line-changes: add support for rendering deleted lines
This adds the IdeLineChangeGutterRenderer:show-line-deletions property.
If set, it will renderer a small red triangle near the line deletion.
It relies on the accuracy of line change tracking in the change monitors.
These are not necessarily perfect today, but this should help us find and
track down any errors with those.