terminal: use theme background for vte background
The widget didn't have any background set, which makes the terminal
have a white background in any theme.

That's a problem when using the dark variant of builder for example.

Fix that applying a style class to the terminal widget, so we can peek
correctly the background color in order to set it manually to the
vte widget.

https://bugzilla.gnome.org/show_bug.cgi?id=762263