configuration: ensure display-name changes get persisted to disk
If we only change the display name, we will not get persisted back to
disk. We don't really want to set the dirty bit just from changing the name
as that would result in a new autogen, so simply emit the changed signal
so that the configuration manager will queue the writeback.