gnome-builder: Tooltips for the greeter view icons
I have created this patch to enhance the user experience on this application, on the greeter view. I have added tooltips to the "New project, Search through projects, Select projects, the checkbox where you choose projects and the cancel searching button". The property is also translatable.
Now the patch is updated with more professional and adequated tooltips.

https://bugzilla.gnome.org/show_bug.cgi?id=755185