build: install headers to $(includedir)/gnome-builder-@VERSION@
To build vala and C plugins out of tree, we need to install the headers.
These are all private API still, and subject to change, so we will install
to a versioned directory based on the GNOME Builder release.

Use pkg-config to get the includes for the version you are using, such as
`pkg-config --cflags --libs libide-1.0`

One might argue that we should drop the -1.0 for libide until we reach
some sort of ABI stability…