vala: add workaround for vala API/ABI break
This checks to see if vala_symbol_get_deprecated() is available. If so,
we use the old version. Otherwise, we use the new style (which is currently
only available in git).

This still doesn't fix the issue with `make distcheck` released tarballs,
since we are still using _SOURCES for *.vala.