egg: add EggEmptyState widget
This is a convenience widget for creating empty-state/initial-state
screens. You can embed a link in the subtitle by escaping the markup
in your UI file. To activate an action, set the uri to

  action://group.name::param

The ::param is optional, but can be used to pass a gvariant parameter to
the action.

For example, to activate the "new-project" action of "app", use:

  action://app.new-project

as the URI for an <a href=""></a> link in the subtitle markup.