sourceview: don't stash GtkTextIter in the private structure
I happened to have missed this during previous review of the mouse-over
click-to-jump feature.

We were stashing GtkTextIter in the private structure and then using them
for comparisons in the future. Obviously, if an edit came in during that
process, we'd have some fun bugs to track down.

The data used for the lookup was also leaked, which is now fixed.

This keeps a couple of marks around to surround the word to be highlighted.
It also uses new marks during the lookup process.

Afterwards, we discard our temp marks, and update the persistent mark
locations.