file-search: update index when new file is created by buffer-manager
Long term, I think we want a better way to update the index when a file
is added to the project. But to do that, we need to introduce some signals
to IdeProject or similar structure (perhaps IdeContext).

The easiest way to fix this bug for 3.20 is to watch for the new file
being created by the buffer manager, since new files in the project tree
get immediately opened.

https://bugzilla.gnome.org/show_bug.cgi?id=760586