panel: rebase IdeLayout and IdeLayoutPane on panel-gtk
Just in time for 3.20, we are landing the panel cleanup. This primarily
allows us to improve things for the dot releases without breaking ABI,
and additionally will fix a bunch of resize warnings introduced by gadgets
in GTK+.

To the user, things should be pretty much the same. However, we've laid
the groundwork to be able to incrementally add panel dragging and
reordering in upcoming releases.

Additionally, we'll be able to add a floating layer to the system in good
time (PnlDockOverlay).