build: take webkit2gtk-4.0 2.12 as a dependency
Various plugins really don't like to be initializing webkit from inside
their compartment, so just take a webkit dependency at the toplevel. This
will make it harder for some systems to build (OSX, etc), but I'm not all
that concerned. It's more important to me that this starts working now.