makecache: parse _VALAFLAGS from automake
Much like our C and C++ parsers, this is a fairly brute force approach.
We fake the VALAC compiler and then work our way through the following
strings. We pretend we are a .c file to determine the target name, which
we can then strip the per-object suffix (foo_la-foo.lo) and resolve to
the .stamp file used for generation (foo_la.stamp).