egg: add EggSimpleLabel
GtkLabel is a lovely piece of engineering, but teaching it to avoid size
requests under the requirements we have is much harder than just rendering
pango layouts ourself.

In particular, it is very important that we avoid a queue_resize() when
changing the contents of the label when possible. We detect this with
the width-chars property and knowning that our content is numeric.

We do perform queue_resize() in a variety of situations, such as the
length of the new string being larger than our width-chars, or when
we drop back under the width-chars size.

This does not currently support baseline due to the tab bar in Builder
not supporting baseline yet.