editor: don't let GtkLabel ruin the performance party
At the top of the editor, we display the current cursor position from
the textview. This has the disastrous side-effect of calling
gtk_widget_queue_resize() every time the label contents changes. Which
as you can imagine, is pretty much every keystroke (unless it was an
accelerator that didn't move the cursor).

When performing the resize, the textview was getting in a situation that
caused it to either do a full redraw or an unnecessary damage.

By using our helper EggSimpleLabel, we can avoid resize cycles in most
circumstances. This results in no more resize events every keypress
while moving with j/k/up/down.