util: add helper to remove GtkTextTag with minimal damage
Normally, removing text tags is done with an optimization to just wipe
the index of all items and damage the whole view. However, this causes
lots of drawing during interactive editing (where showing and removing
error squiggles is common).

To work around this, we just do the removal of items manually which
damages far less of the screen real-estate.