jedi: don't hardcode the path of gir files
FreeBSD and OpenBSD don't install GNOME and most other software
packages with /usr prefix. The installation prefix can be changed by
users, but it must not be /usr to avoid breaking the system. It is
usually better to rely on environment variables and gobject-introspection
itself to provide the path.

This change also allows reading gir files from multiple prefixes,
enabling the use of gir files installed in a JHBuild prefix. gir files
located in different directories but with the same name are not processed
twice to avoid inserting duplicate items into the database.

https://bugzilla.gnome.org/show_bug.cgi?id=763263