egg-simple-label: replace alloca with C99 VLA
alloca is a non-standard feature. As we already requires a compiler
supporting C11, using the standard variable-length array seems to be
better. Although VLA is an optional feature in C11, both gcc and clang
support it for a long time.

This commit fixes the build on FreeBSD, which puts alloca in stdlib.h
instead of alloca.h.

https://bugzilla.gnome.org/show_bug.cgi?id=765441