search: avoid g_signal_handlers_disconnect_by_func()
This hits the slow path on disconnection, and we go through search
contexts on every keypress. Worth avoiding all the handlers_find()
overhead in gsignal.c

Instead, we just track the handler id and disconnect it directly.

This also ensures we disconnect the two signals we left lingering before.
It wasn't a big deal since the contexts were always immediately released
and we used connect_object(), but better to be correct.