prefs: prevent deselection of radio in option group
If we have multiple options for a radio, we might end up deselecting our
current selection (as the changed event doesn't propagate since we are
already trying to update).

This is a quick workaround that just does a final pass to read the new
value and reapply the check if necessary.

https://bugzilla.gnome.org/show_bug.cgi?id=766086