autotools: log process spawn in main thread
This was originally intended to ever get called from the main thread, but
due to simplifying some of the complexity got pushed into a worker thread.

Make sure we log the process spawning via the main thread, so that the
event dispatched to gtk+ is on the right thread.

We should probably add some helpers to IdeBuildResult to handle running
from other threads and propagate those log entries to the main thread.
But that can wait until later, as this fixes the issue.