realtime-graphs: try harder to not jump when resizing
We were resetting our x_offset when resizing the window, so things would
appear to jump backwards. This tries to persist the offset in time and
only apply the alloc.width multiplication when it comes time to render.

This allows us to keep the x_offset between resizes, and therefore endup
on the right offset even when destroying the surface.