editor: avoid async spinloop
We could get in a state if the view was closed while we were loading it
that we would try again but fail. And then continue to do that since the
buffer was already open (but no longer necessary to attach a view to it).

This just prevents us from trying a third time in the retry case.