clang: use copy of IdeFile for cache key
While I don't really like our use of IdeFile in general, its a bit hard
to remove from the codebase today. It can hold onto file-settings which
we'd prefer to have disposed sooner rather than later. So use an IdeFile
copy which won't have any of that attached to it.