buffer: Allow buffers to be loaded without displaying them
In order for :vsplit <file> to work, the buffer has to be loaded in a
separate IdeLayoutStack rather than the normal way of adding an
IdeEditorView to the existing stack. This commit adds a parameter to
ide_workbench_open_files_async and the functions further down the call
stack so that buffers can be loaded without immediately displaying them
to the user. Instead, a callback will be used to load the buffer when
the split is created.

https://bugzilla.gnome.org/show_bug.cgi?id=767540