build: reorganize libide
The libide/ directory was getting a bit out of control. The number of files
in the directory was simply too much to make it easy for newcomers to
find what they are looking for.

This breaks things up into a bunch of, somewhat logical, directories.

We no longer use a bunch of crazy include paths with -I, and instead
include the directory name when including as such:

  #include "buffers/ide-buffer.h"

This will hopefully make it easier to navigate, and basically the same to
open files if you use fuzzy search to open files.