sourceview: port to gtk_source_search_context_forward2 and variants
The previous functions are now deprecated, and the new variant includes
a wrapping parameter. We might be able to simplify these functions a bit,
but I need to look into it more. I think the wrapping simply notifies
if we wrapped around to the start of the buffer, not if we are passing
a particular marker that we care about (which is what we need).