headerbar: add hack for center right/left alignment
To implement our designs propertly would require GtkHeaderBar to support
hexpand for start/end children. Instead, we can work around it for now
using a center child and opposite packing.

It's not perfect, since it will require even allocation on left/right, but
it allows us to move forward for now.