headerbar: add helper to get IdeOmniBar
The OmniBar is pretty much what it says, the joining of a bunch of things
into a single widget. Getting the density of information we want definitely
can result in non-clean abstractions.

However, if we can come up with a good way to have extension points in the
omnibar, we might just be able to clean this up.