listbox: add GtkListBox subclass that can recycle rows
This isn't the ideal version we want where we can limit the number of
inflated rows based on the viewport, but it does help us in situations
where we have to recreate rows a lot (like when updating search results).

The goal here is to replace the search results with something like this so
we get less flicker on wayland when repopulating search results.

If we don't add any more complexity to our search results (unlikely), we
could of course just switch it to TreeView with minimal effort.