omnibar: add build/rebuild/clean actions to popover
We need somewhere to put this so we don't lose the feature, and putting
it in the omnibar popover seems prudent (as that is where other information
lives). This may not be our ideal styling, but I think it looks good
enough to help us land this branch sooner than later.