editor: register accelerator for editor perspective
I'm not sure if this one will live long term, but we discussed the idea
of having <alt>1...n for some typical perspectives as they land.

Still a bit unsure of it though.

What I wish we could do is register <alt>F1..n for these perspectives,
but many of those are taken by things like help, or the shell.