posix: add helpers to expand/collapse user paths
Generally, using ~ in paths inside the program is tedious, but there are a
few use cases where we want to support that. (Such as the path used in the
create project dialog).

This is a simple helper that uses the POSIX wordexp(3) function to expand
simple shell features. It does not, however, perform command substitution.