git: use EggFileChooserEntry
While this removes the ability to set the exact directory name (and instead
defaults to the name of the remote project) it is what users want most of
the time.

We might want to detect if the directory does not exist, and in that case
we can put it into the exact directory that was provided.