app: add handler to open files from command line
This fixes support for `gnome-builder path/to/some/file.c`. It will try to
locate the project (which might be path/configure.ac) and load it first,
the the file within that workbench.

However, Builder currently does a lot of setup work during startup (which
should be improved in followup commits). So this will have a bit of a delay
while starting up and is not as fast as say, running `gedit foo.c`.

In particular, we should be able to avoid scanning for recent files in this
situation.