editor-frame: Only enable replace/replace-all at appropriate times
When the search entry field is empty, the replace and replace-all
actions don't make sense, so the buttons should be insensitive.
Similarly, if no search occurrence is selected, the replace action
doesn't make sense. This commit sets up callbacks to try to ensure the
buttons are only sensitive (clickable) at appropriate times.

https://bugzilla.gnome.org/show_bug.cgi?id=765635