command-bar: Allow command bar actions to steal focus
Currently, gb_command_bar_hide grabs focus on whatever was focused
before the command bar popped up, or on the workbench. But if a command
bar action causes another widget to appear, we want to leave that in
focus. For example, the substitute action with the 'c' option will have
that behavior (:%s/this/that/gc).

https://bugzilla.gnome.org/show_bug.cgi?id=765635