command-bar: Implement the 'c' option for vim substitute commands
In vim, if the 'c' option is specified at the end of a substitute
command (like :%s/this/that/gc), the user is prompted to confirm or deny
replacement on each occurrence of the search term. This commit makes
Builder recognize that option and respond by opening the replace UI with
the search entry and replace entry fields filled in.

Because the search_revealer animation doesn't happen immediately (it
waits for the frame clock to increment), we can't directly trigger the
"next-search-result" action in ide_editor_frame_actions_replace_confirm.
The selection would be cleared once the search_entry widget is mapped.
So instead we setup a callback on the buffer's has-selection property,
and use that to trigger next-search-result if there's a pending
replace-confirm action.

https://bugzilla.gnome.org/show_bug.cgi?id=765635