greeter: use SELECT_FOLDER when appropriate
If we have a file filter that is of type inode/directory, then we really
need SELECT_FOLDER so that the user can select the directory instead of
diving into the directory. This will break things that have content types
mixed with inode/directory, but I think that is okay.