vim: Check for open buffer in :bnext and :bprev commands
The vim commands :bnext and :bprev let the user switch to another
buffer. If those commands are executed without any buffers open,
ide_widget_action() fails and emits a warning. This commit checks if any
buffers are open, and if not the command gracefully gives up. This
matches vim's behavior.

https://bugzilla.gnome.org/show_bug.cgi?id=768356