indenter-c: Allow multi-line conditions to end a line with ')'
Currently the indenter assumes that a condition is complete if one of
its lines ends with a matched ')'. If for some reason you want to end a
line with ')' but not end the condition, the auto indenting would be
wrong. This commit fixes that by checking for an unmatched '(' when a
line ends with ')'.

https://bugzilla.gnome.org/show_bug.cgi?id=768459