editor-frame: Prevent search position label flickering during replace
In ide_editor_frame_actions_replace, when the replace is executed the
deletion and insertion in the buffer cause the search position label
(e.g. "1 of 29") to be updated, but at that time the occurrences-count
is unknown since that part of the buffer needs to be re-scanned, so the
label is set to NULL. Only once the scan finishes can the corect label
be set. So this commit prevents the search position label from being
updated when the cursor moves during the replacement, which prevents
ugly flickering for the user.

https://bugzilla.gnome.org/show_bug.cgi?id=768534