style-schemes: Use the style scheme for the search shadow color
The search shadow color is the color of everything in the buffer other
than search results. The setting for it is not currently turned on in
Builder so you have to set IdeSourceView's show-search-shadow property
to TRUE to see it. This commit moves the color of the shadow to the
style scheme so it is more easily configurable. It also renames the
variable used for the search-match style object to make the code more
readable.

https://bugzilla.gnome.org/show_bug.cgi?id=768580